moved actionTest route into routes directory
